added extension method to simplify copying
